+++ tracker/src/app.c
@@ -1,5 +1,6 @@
 #include <string.h>
 #include "corelib_gfx.h"
+#include "corelib_font.h"
 #include "common.h"
 #include "audio_manager.h"
 #include "app.h"